Image Crop SDK example
Added an example of how to crop images with opencv. Intended to provide
an example for Danny Data who wants to do arbitrary image pre-processing
without commiting to either developing fairly complex user facing
surface area or very limited and not too useful functionality. One
concern is lack of discoverability. Considered putting a how-to page
into the SDK docs but it would be odd to do that I think.